functions/extract-summary/requirements.txt
newspaper3k==0.2.*
nltk==3.5.*
boilerpy3==1.0.*
google-cloud-secret-manager==1.0.*
